+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 2024/10/19
+### 
+
+### Added
+- pg_search gem
+    - Gemfile
+    - Gemfile.lock
+
+### Changed
+- readme
+    - referenced pg search as a dependency
+- associated_attrs_controller.rb
+    - search method to check DB against assocaiated_attrs -> attr_title
+- associated_attrs/index.html.erb
+    - search bar implemented
+- routes.rb
+    - updated to make pg_search work
+
+## [Unreleased] 2024/10/19
 ### Other seed files now seed the db with '{#seed_file}' into associated_endpoint field
 
 ### Added